\name{Loregic-package}
\alias{Loregic-package}
\alias{Loregic}
\docType{package}
\title{
Loregic: A method to characterize the cooperative logic of regulatory factors}
\description{
Loregic is a computational method inputting gene expression and regulatory network data, to characterize the cooperativity of regulatory factors. Loregic uses all 16 possible two-input-one-output logic gates (e.g. AND or XOR) to describe triplets of two factors regulating a common target. It attempts to find the gate that best matches each triplet’s observed gene expression across many conditions.}
\details{
\tabular{ll}{
Package: \tab Loregic\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-09-08\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Daifeng Wang, Mark Gerstein
Maintainer: Mark Gerstein <pi@gersteinlab.org>
}
\references{
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ logic gate, gene regulatory network, regulatory cooperativity }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
